requirements_s3.txt
# Dependencias para el script de subida a S3
boto3>=1.28.0
python-dotenv>=1.0.0
botocore>=1.31.0
pydantic>=2.10.7
pydantic-settings>=2.10.7